Add volume and snapshot tracking in global configuration file

+++ commands.go
@@ -14,6 +14,10 @@ func getDriverRoot(root, driverName string) string {
 	return filepath.Join(root, driverName) + "/"
 }
 
+func getConfigFileName(root string) string {
+	return filepath.Join(root, CONFIGFILE)
+}
+
 func doInitialize(root, driverName string, driverOpts map[string]string) error {
 	utils.MkdirIfNotExists(root)
 	log.Debug("Config root is ", root)
@@ -27,10 +31,11 @@ func doInitialize(root, driverName string, driverOpts map[string]string) error {
 		return err
 	}
 
-	configFileName := filepath.Join(root, CONFIGFILE)
+	configFileName := getConfigFileName(root)
 	config := Config{
-		Root:   root,
-		Driver: driverName,
+		Root:    root,
+		Driver:  driverName,
+		Volumes: make(map[string]Volume),
 	}
 	err = utils.SaveConfig(configFileName, &config)
 	return err
@@ -44,14 +49,27 @@ func doInfo(config *Config, driver drivers.Driver) error {
 
 func doVolumeCreate(config *Config, driver drivers.Driver, size uint64) error {
 	uuid := uuid.New()
-	base := ""
-	err := driver.CreateVolume(uuid, base, size)
+	base := "" //Doesn't support base for now
+	if err := driver.CreateVolume(uuid, base, size); err != nil {
+		return err
+	}
+	log.Debug("Created volume using ", config.Driver)
+	config.Volumes[uuid] = Volume{
+		Base:      base,
+		Size:      size,
+		Snapshots: make(map[string]bool),
+	}
+	err := utils.SaveConfig(getConfigFileName(config.Root), config)
 	return err
 }
 
 func doVolumeDelete(config *Config, driver drivers.Driver, uuid string) error {
-	log.Debug("Deleting volume using ", config.Driver)
-	err := driver.DeleteVolume(uuid)
+	if err := driver.DeleteVolume(uuid); err != nil {
+		return err
+	}
+	log.Debug("Deleted volume using ", config.Driver)
+	delete(config.Volumes, uuid)
+	err := utils.SaveConfig(getConfigFileName(config.Root), config)
 	return err
 }
 
@@ -65,13 +83,32 @@ func doVolumeList(config *Config, driver drivers.Driver) error {
 }
 
 func doSnapshotCreate(config *Config, driver drivers.Driver, volumeUUID string) error {
+	if _, exists := config.Volumes[volumeUUID]; !exists {
+		return fmt.Errorf("volume %v doesn't exist", volumeUUID)
+	}
 	uuid := uuid.New()
-	err := driver.CreateSnapshot(uuid, volumeUUID)
+	if err := driver.CreateSnapshot(uuid, volumeUUID); err != nil {
+		return err
+	}
+	log.Debugf("Created snapshot %v of volume %v using %v\n", uuid, volumeUUID, config.Driver)
+
+	config.Volumes[volumeUUID].Snapshots[uuid] = true
+	err := utils.SaveConfig(getConfigFileName(config.Root), config)
 	return err
 }
 
 func doSnapshotDelete(config *Config, driver drivers.Driver, uuid, volumeUUID string) error {
+	if _, exists := config.Volumes[volumeUUID]; !exists {
+		return fmt.Errorf("volume %v doesn't exist", volumeUUID)
+	}
+	if _, exists := config.Volumes[volumeUUID].Snapshots[uuid]; !exists {
+		return fmt.Errorf("snapshot %v of volume %v doesn't exist", uuid, volumeUUID)
+	}
 	err := driver.DeleteSnapshot(uuid, volumeUUID)
+	log.Debugf("Deleted snapshot %v of volume %v using %v\n", uuid, volumeUUID, config.Driver)
+
+	delete(config.Volumes[volumeUUID].Snapshots, uuid)
+	err = utils.SaveConfig(getConfigFileName(config.Root), config)
 	return err
 }
 
@@ -103,9 +140,17 @@ func doBlockStoreDeregister(config *Config, kind, id string) error {
 }
 
 func doBlockStoreAdd(config *Config, blockstoreId, volumeId string) error {
+	if _, exists := config.Volumes[volumeId]; !exists {
+		return fmt.Errorf("volume %v doesn't exist", volumeId)
+	}
+
 	return blockstores.AddVolume(getBlockStoreRoot(config.Root), blockstoreId, volumeId)
 }
 
 func doBlockStoreRemove(config *Config, blockstoreId, volumeId string) error {
+	if _, exists := config.Volumes[volumeId]; !exists {
+		return fmt.Errorf("volume %v doesn't exist", volumeId)
+	}
+
 	return blockstores.RemoveVolume(getBlockStoreRoot(config.Root), blockstoreId, volumeId)
 }
